--- c10/util/Metaprogramming.cpp
@@ -1,1 +1,0 @@
-#include <c10/util/Metaprogramming.h>
