+++ test/cpp/aoti_abi_check/test_typetraits.cpp
@@ -1,7 +1,7 @@
-#include <c10/util/TypeTraits.h>
+#include <torch/headeronly/util/TypeTraits.h>
 #include <gtest/gtest.h>
 
-using namespace c10::guts;
+using namespace torch::headeronly::guts;
 
 // NOLINTBEGIN(modernize-unary-static-assert)
 namespace {
