+++ torch/header_only_apis.txt
@@ -57,6 +57,39 @@ _isnan
 Philox4_32
 randn
 
+# torch/headeronly/Metaprogramming.h
+function_traits
+make_function_traits_t
+infer_function_traits_t
+tuple_elements
+tuple_take
+tuple_map
+
+# torch/headeronly/TypeList.h
+# false_t => tested through from_tuple_t and various other APIs
+typelist
+size
+from_tuple_t
+to_tuple_t
+concat_t
+filter_t
+count_if
+map_t
+head_t
+head_with_default_t
+reverse_t
+find_if
+contains
+take_t
+drop_t
+drop_if_nonempty_t
+
+# torch/headeronly/TypeTraits.h
+is_equality_comparable
+is_function_type
+is_instantiation_of
+is_stateless_lambda
+
 # ATen/cpu/vec/vec.h
 Vectorized
 clamp_min
